refactor(core): split azure storage backend helpers
