+++ crates/surge-core/src/storage/azure/requests.rs
@@ -1,32 +1,16 @@
-//! Azure Blob Storage backend with SharedKey authentication.
-
-use async_trait::async_trait;
 use std::path::Path;
 
-use base64::Engine;
-use base64::engine::general_purpose::STANDARD as BASE64;
-use chrono::Utc;
+use async_trait::async_trait;
 use reqwest::Client;
-use tracing::{debug, trace};
+use tracing::debug;
 
 use crate::context::StorageConfig;
-use crate::crypto::hmac_sha256::hmac_sha256;
 use crate::error::{Result, SurgeError};
-use crate::storage::{ListEntry, ListResult, ObjectInfo, StorageBackend, TransferProgress, download_response_to_file};
+use crate::storage::{ListResult, ObjectInfo, StorageBackend, TransferProgress, download_response_to_file};
 
-/// Azure Blob Storage REST API version.
-const AZURE_API_VERSION: &str = "2024-08-04";
-
-/// Azure Blob Storage backend using SharedKey authentication.
-pub struct AzureBlobBackend {
-    client: Client,
-    account: String,
-    /// Base64-decoded account key (raw bytes) for HMAC signing.
-    key_bytes: Vec<u8>,
-    container: String,
-    endpoint: String,
-    prefix: String,
-}
+use super::AzureBlobBackend;
+use super::auth::decode_account_key;
+use super::listing::parse_azure_list_blobs_xml;
 
 impl AzureBlobBackend {
     /// Create a new Azure Blob Storage backend from configuration.
@@ -59,14 +43,6 @@ impl AzureBlobBackend {
             ));
         }
 
-        let key_bytes = if account_key.is_empty() {
-            Vec::new()
-        } else {
-            BASE64
-                .decode(&account_key)
-                .map_err(|e| SurgeError::Config(format!("Azure account key is not valid base64: {e}")))?
-        };
-
         let endpoint = if config.endpoint.is_empty() {
             format!("https://{account_name}.blob.core.windows.net")
         } else {
@@ -87,27 +63,11 @@ impl AzureBlobBackend {
         Ok(Self {
             client,
             account: account_name,
-            key_bytes,
+            key_bytes: decode_account_key(&account_key)?,
             container: config.bucket.clone(),
             endpoint,
             prefix: config.prefix.clone(),
         })
-    }
-
-    /// Returns `true` when credentials are available for signing requests.
-    fn has_credentials(&self) -> bool {
-        !self.key_bytes.is_empty()
-    }
-
-    /// Return an error when a write is attempted without credentials.
-    fn require_credentials(&self, operation: &str) -> Result<()> {
-        if self.has_credentials() {
-            Ok(())
-        } else {
-            Err(SurgeError::Config(format!(
-                "Azure Blob {operation} requires account credentials"
-            )))
-        }
     }
 
     /// Build the full blob name, prepending the configured prefix.
@@ -132,124 +92,6 @@ impl AzureBlobBackend {
     /// Build the container-level URL (for list operations).
     fn container_url(&self) -> String {
         format!("{}/{}", self.endpoint.trim_end_matches('/'), self.container)
-    }
-
-    /// Produce an RFC 1123 date string (e.g. `Mon, 02 Mar 2026 12:34:56 GMT`).
-    fn rfc1123_date(now: &chrono::DateTime<Utc>) -> String {
-        now.format("%a, %d %b %Y %H:%M:%S GMT").to_string()
-    }
-
-    /// Sign a request using SharedKey authentication and return the headers
-    /// that must be attached to the request.
-    ///
-    /// The string-to-sign format for Blob service:
-    /// ```text
-    /// VERB\n
-    /// Content-Encoding\n
-    /// Content-Language\n
-    /// Content-Length\n
-    /// Content-MD5\n
-    /// Content-Type\n
-    /// Date\n
-    /// If-Modified-Since\n
-    /// If-Match\n
-    /// If-None-Match\n
-    /// If-Unmodified-Since\n
-    /// Range\n
-    /// CanonicalizedHeaders\n
-    /// CanonicalizedResource
-    /// ```
-    #[allow(clippy::too_many_arguments)]
-    fn sign_request(
-        &self,
-        method: &str,
-        resource_path: &str,
-        query_params: &[(String, String)],
-        content_length: Option<usize>,
-        content_type: &str,
-        extra_headers: &[(String, String)],
-    ) -> Vec<(String, String)> {
-        let now = Utc::now();
-        let date = Self::rfc1123_date(&now);
-
-        // Collect all x-ms-* headers (sorted).
-        let mut ms_headers: Vec<(String, String)> = vec![
-            ("x-ms-date".to_string(), date.clone()),
-            ("x-ms-version".to_string(), AZURE_API_VERSION.to_string()),
-        ];
-        for (k, v) in extra_headers {
-            if k.starts_with("x-ms-") {
-                ms_headers.push((k.clone(), v.clone()));
-            }
-        }
-        ms_headers.sort_by(|a, b| a.0.cmp(&b.0));
-        // Deduplicate on key, keeping later values.
-        ms_headers.dedup_by(|a, b| {
-            if a.0 == b.0 {
-                // Keep value from `b` (the earlier one after sort is stable).
-                true
-            } else {
-                false
-            }
-        });
-
-        let canonicalized_headers = ms_headers
-            .iter()
-            .map(|(k, v)| format!("{k}:{v}"))
-            .collect::<Vec<_>>()
-            .join("\n");
-
-        // Canonicalized resource: /account/container/blob?\ncomp:list\n...
-        let mut canonicalized_resource = format!("/{}{}", self.account, resource_path);
-        if !query_params.is_empty() {
-            let mut sorted_params = query_params.to_vec();
-            sorted_params.sort_by(|a, b| a.0.cmp(&b.0));
-            for (k, v) in &sorted_params {
-                canonicalized_resource.push('\n');
-                canonicalized_resource.push_str(k);
-                canonicalized_resource.push(':');
-                canonicalized_resource.push_str(v);
-            }
-        }
-
-        // Content-Length: omit for 0 or absent.
-        let content_length_str = match content_length {
-            Some(n) if n > 0 => n.to_string(),
-            _ => String::new(),
-        };
-
-        // String to sign.
-        let string_to_sign = format!(
-            "{method}\n\
-             \n\
-             \n\
-             {content_length_str}\n\
-             \n\
-             {content_type}\n\
-             \n\
-             \n\
-             \n\
-             \n\
-             \n\
-             \n\
-             {canonicalized_headers}\n\
-             {canonicalized_resource}"
-        );
-
-        trace!(string_to_sign = %string_to_sign, "Azure string to sign");
-
-        // HMAC-SHA256 with the account key.
-        let signature = BASE64.encode(hmac_sha256(&self.key_bytes, string_to_sign.as_bytes()));
-
-        let authorization = format!("SharedKey {}:{signature}", self.account);
-
-        let mut headers: Vec<(String, String)> = ms_headers;
-        headers.push(("Authorization".to_string(), authorization));
-        if !content_type.is_empty() {
-            headers.push(("Content-Type".to_string(), content_type.to_string()));
-        }
-
-        headers
     }
 
     /// Map an HTTP response status to a `SurgeError`.
@@ -392,7 +234,6 @@ impl StorageBackend for AzureBlobBackend {
 
         let resp = req.send().await?;
         let status = resp.status();
-        // Azure returns 202 Accepted for deletes, and 404 if already gone.
         if !status.is_success() && status != reqwest::StatusCode::ACCEPTED && status != reqwest::StatusCode::NOT_FOUND {
             let body = resp.text().await.unwrap_or_default();
             Self::check_response_status(status, &full_key, &body)?;
@@ -406,7 +247,6 @@ impl StorageBackend for AzureBlobBackend {
         let full_prefix = self.full_key(prefix);
         let container_url = self.container_url();
 
-        // Query parameters for List Blobs.
         let mut query_params: Vec<(String, String)> = vec![
             ("comp".to_string(), "list".to_string()),
             ("restype".to_string(), "container".to_string()),
@@ -501,104 +341,4 @@ impl StorageBackend for AzureBlobBackend {
         }
         Ok(())
     }
-}
-
-// ---------------------------------------------------------------------------
-// XML parsing for Azure List Blobs response
-// ---------------------------------------------------------------------------
-
-/// Parse an Azure List Blobs XML response into a `ListResult`.
-///
-/// Azure response structure:
-/// ```xml
-/// <EnumerationResults>
-///   <Blobs>
-///     <Blob>
-///       <Name>key</Name>
-///       <Properties>
-///         <Content-Length>123</Content-Length>
-///       </Properties>
-///     </Blob>
-///   </Blobs>
-///   <NextMarker>...</NextMarker>
-/// </EnumerationResults>
-/// ```
-fn parse_azure_list_blobs_xml(xml: &str) -> Result<ListResult> {
-    use quick_xml::Reader;
-    use quick_xml::events::Event;
-
-    let mut reader = Reader::from_str(xml);
-    let mut buf = Vec::new();
-    let mut entries = Vec::new();
-    let mut next_marker: Option<String> = None;
-
-    let mut in_blob = false;
-    let mut in_properties = false;
-    let mut current_name: Option<String> = None;
-    let mut current_size: Option<i64> = None;
-    let mut current_tag = String::new();
-
-    loop {
-        match reader.read_event_into(&mut buf) {
-            Ok(Event::Start(ref e)) => {
-                let tag = String::from_utf8_lossy(e.name().as_ref()).to_string();
-                match tag.as_str() {
-                    "Blob" => {
-                        in_blob = true;
-                        current_name = None;
-                        current_size = None;
-                    }
-                    "Properties" if in_blob => {
-                        in_properties = true;
-                    }
-                    _ => {}
-                }
-                current_tag = tag;
-            }
-            Ok(Event::End(ref e)) => {
-                let tag = String::from_utf8_lossy(e.name().as_ref()).to_string();
-                match tag.as_str() {
-                    "Blob" => {
-                        if let Some(name) = current_name.take() {
-                            entries.push(ListEntry {
-                                key: name,
-                                size: current_size.unwrap_or(0),
-                            });
-                        }
-                        in_blob = false;
-                        in_properties = false;
-                    }
-                    "Properties" => {
-                        in_properties = false;
-                    }
-                    _ => {}
-                }
-                current_tag.clear();
-            }
-            Ok(Event::Text(ref e)) => {
-                let text = String::from_utf8_lossy(e.as_ref()).to_string();
-                if in_blob && !in_properties && current_tag == "Name" {
-                    current_name = Some(text);
-                } else if in_properties && current_tag == "Content-Length" {
-                    current_size = text.parse::<i64>().ok();
-                } else if !in_blob && current_tag == "NextMarker" && !text.is_empty() {
-                    next_marker = Some(text);
-                }
-            }
-            Ok(Event::Eof) => break,
-            Err(e) => {
-                return Err(SurgeError::Storage(format!("Failed to parse Azure list response: {e}")));
-            }
-            _ => {}
-        }
-        buf.clear();
-    }
-
-    let is_truncated = next_marker.is_some();
-    debug!(count = entries.len(), is_truncated, "Azure LIST parsed");
-    Ok(ListResult {
-        entries,
-        next_marker,
-        is_truncated,
-    })
 }
