+++ docs/architecture/cleanup-plan.md
@@ -35,36 +35,36 @@ These PRs are already merged:
 - `#63` `refactor(core): split release restore planning and artifact recovery`
 - `#64` `refactor(core): split shortcut installation by platform`
 - `#65` `refactor(core): split manifest module responsibilities`
+- `#66` `refactor(core): split install module responsibilities`
 
 ## Active Phase
 
-### `refactor/install-phase-1`
+### `refactor/storage-azure-phase-1`
 
 Current goal:
 
-- split [`crates/surge-core/src/install/mod.rs`](../../crates/surge-core/src/install/mod.rs)
+- split [`crates/surge-core/src/storage/azure/mod.rs`](../../crates/surge-core/src/storage/azure/mod.rs)
   into:
-  - `install/runtime_manifest.rs`
-  - `install/activation.rs`
-  - `install/persistent_assets.rs`
-  - `install/launch.rs`
+  - `storage/azure/auth.rs`
+  - `storage/azure/requests.rs`
+  - `storage/azure/listing.rs`
 
 Current checkpoint:
 
 - the leaf modules have been created
-- the root module has been reduced to install-facing types, entrypoints, and tests
+- the root module has been reduced to the backend type and module wiring
 - targeted compile of `surge-core` passes
-- focused `install` tests pass
+- focused `storage::azure` tests pass
 - focused `surge-core` clippy passes
-- the install baseline entry has been removed
+- the azure baseline entry has been removed
 - the full pre-push suite passes on the branch
 
 Exit criteria:
 
-- `cargo test -p surge-core install` passes
+- `cargo test -p surge-core storage::azure` passes
 - `cargo clippy -p surge-core --all-targets --all-features -- -D warnings -W clippy::pedantic` passes
 - `./scripts/check-maintainability.sh` reports the file below the target so the
-  install baseline entry can be removed
+  azure baseline entry can be removed
 - the full pre-push suite passes
 - the PR is merged with squash, local cleanup is done, and merged-`main` CI is green
 
@@ -72,14 +72,13 @@ Exit criteria:
 
 These are the remaining planned PRs from the original Rust-first campaign.
 
-### 1. `refactor/install-phase-1`
+### 1. `refactor/storage-azure-phase-1`
 
-- split [`crates/surge-core/src/install/mod.rs`](../../crates/surge-core/src/install/mod.rs)
+- split [`crates/surge-core/src/storage/azure/mod.rs`](../../crates/surge-core/src/storage/azure/mod.rs)
   into focused modules for:
-  - runtime manifest handling
-  - activation and snapshot pruning
-  - persistent asset copying and validation
-  - launch and post-install lifecycle helpers
+  - shared-key signing and credential checks
+  - request construction and backend operations
+  - list-response XML parsing
 
 ### 2. `refactor/maintainability-phase-2`
 
@@ -103,7 +102,6 @@ be decomposed to fully retire the baseline.
 ### Core surfaces
 
 - [`crates/surge-core/src/releases/delta.rs`](../../crates/surge-core/src/releases/delta.rs)
-- [`crates/surge-core/src/storage/azure.rs`](../../crates/surge-core/src/storage/azure.rs)
 - [`crates/surge-core/src/storage/gcs.rs`](../../crates/surge-core/src/storage/gcs.rs)
 
 ### Bench debt
